feat: reporter adapter streaming ssr
