+++ .changeset/twenty-panthers-happen.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/runtime': minor
+---
+
+feat: reporter adapter streaming ssr
+feat: reporter 适配 streaming ssr
