+++ packages/runtime/plugin-runtime/src/ssr/index.node.tsx
@@ -5,6 +5,7 @@ import type {
 } from './serverRender/types';
 import { formatServer } from './utils';
 import render from './serverRender';
+import { createSSRTracker } from './serverRender/tracker';
 
 export const ssr = (config: SSRPluginConfig = {}): Plugin => ({
   name: '@modern-js/plugin-ssr',
@@ -23,6 +24,7 @@ export const ssr = (config: SSRPluginConfig = {}): Plugin => ({
 
         context.ssrContext!.request = formatServer(request);
         context.ssrContext!.mode = config.mode;
+        context.ssrContext!.tracker = createSSRTracker(context.ssrContext!);
 
         if (!context.ssrContext!.htmlModifiers) {
           context.ssrContext!.htmlModifiers = [];
