+++ packages/runtime/plugin-runtime/src/ssr/serverRender/renderToStream/index.ts
@@ -3,6 +3,7 @@ import { run } from '@modern-js/utils/runtime-node';
 import { time } from '@modern-js/utils/universal/time';
 import { PreRender } from '../../react/prerender';
 import { ServerRenderOptions } from '../types';
+import { SSRTimings } from '../tracker';
 import renderToPipe from './renderToPipe';
 
 export const render = ({ App, context }: ServerRenderOptions) => {
@@ -21,6 +22,8 @@ export const render = ({ App, context }: ServerRenderOptions) => {
       }),
     });
 
+    const { tracker } = ssrContext;
+
     const pipe = renderToPipe(rootElement, context, {
       onShellReady() {
         // set cacheConfig
@@ -28,12 +31,14 @@ export const render = ({ App, context }: ServerRenderOptions) => {
         if (cacheConfig) {
           ssrContext.cacheConfig = cacheConfig;
         }
+
+        const cost = end();
+        tracker.trackTiming(SSRTimings.SSR_RENDER_SHELL, cost);
       },
       onAllReady() {
         // calculate streaming ssr cost
         const cost = end();
-        ssrContext.logger.debug('App Render To HTML cost = %d ms', cost);
-        ssrContext.metrics.emitTimer('app.render.html.cost', cost);
+        tracker.trackTiming(SSRTimings.SSR_RENDER_TOTAL, cost);
       },
     });
     return pipe;
