+++ packages/runtime/plugin-runtime/src/ssr/serverRender/types.ts
@@ -1,8 +1,19 @@
 import { ServerUserConfig } from '@modern-js/app-tools';
+import type { BaseSSRServerContext } from '@modern-js/types';
 import type { RuntimeContext } from '../../core';
 import { RenderLevel } from './renderToString/type';
+import type { BuildHtmlCb } from './renderToString/buildHtml';
+import type { SSRTracker } from './tracker';
 
-export type { SSRServerContext } from './renderToString/type';
+export type SSRServerContext = BaseSSRServerContext & {
+  request: BaseSSRServerContext['request'] & {
+    userAgent: string;
+    cookie: string;
+    cookieMap: Record<string, string>;
+  };
+  htmlModifiers: BuildHtmlCb[];
+  tracker: SSRTracker;
+};
 export type ModernSSRReactComponent = React.ComponentType<any>;
 export { RuntimeContext, RenderLevel };
