+++ tests/integration/ssr/fixtures/streaming/src/routes/error/page.tsx
@@ -2,14 +2,14 @@ import { Await, useAsyncError, useLoaderData } from '@modern-js/runtime/router';
 import { Suspense } from 'react';
 
 export default function Page() {
-  const data = useLoaderData();
+  const data = useLoaderData() as any;
 
   return (
     <div>
       Error page
       <Suspense fallback={<div>loading ...</div>}>
         <Await resolve={data.data} errorElement={<ErrorElement />}>
-          {(data: any) => {
+          {_ => {
             return <div>never shown</div>;
           }}
         </Await>
